// cypress/support/types/User.ts
export type User = {
  firstName: string;
  lastName: string;
  email: string;
  gender: string;
  mobile: string;
  subjects: string;
  hobbies: string;
  address: string;
  pic: string;
  state: string;
  city: string;
  dateOfBirthDay: string;
  dateOfBirthMonth: string;
  dateOfBirthYear: string;
};